refactor(DL-16084): Addressed PlatOps comments
